Task_7/requirements.txt
tensorflow==2.8.0
numpy==1.19.5
Pillow==8.4.0